admin: Move SetCredentials from Service to Generic
Setting credentials doesn't belong to service management API
anymore.
